requirements.txt
PyMuPDF==1.23.26
pdfplumber==0.11.4
pytesseract==0.3.10
Pillow==10.2.0
langdetect==1.0.9